+++ pos_multi_session/__manifest__.py
@@ -1,10 +1,17 @@
 # -*- coding: utf-8 -*-
+# Copyright 2017-2018 Ivan Yelizariev <https://it-projects.info/team/yelizariev>
+# Copyright 2017 Kolushov Alexandr <https://it-projects.info/team/KolushovAlexandr>
+# Copyright 2017 Ilmir Karamov <https://it-projects.info/team/ilmir-k>
+# Copyright 2017-2018 Dinar Gabbasov <https://it-projects.info/team/GabbasovDinar>
+# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl.html).
+
 {
     "name": """Sync POS orders across multiple sessions""",
     "summary": """Use multiple POS for handling orders""",
     "category": "Point Of Sale",
+    "live_test_url": 'http://apps.it-projects.info/shop/product/pos-multi-session?version=10.0',
     "images": ["images/pos-multi-session.png"],
-    "version": "3.1.0",
+    "version": "10.0.4.1.0",
     "application": False,
 
     "author": "IT-Projects LLC, Ivan Yelizariev",
@@ -16,10 +23,11 @@
 
     "depends": [
         "pos_disable_payment",
-        "pos_longpolling",
+        "pos_multi_session_sync"
     ],
     "external_dependencies": {"python": [], "bin": []},
     "data": [
+        "data/pos_multi_session_data.xml",
         "security/ir.model.access.csv",
         "views/pos_multi_session_views.xml",
         "multi_session_view.xml"
@@ -36,5 +44,19 @@
     "post_init_hook": None,
 
     "auto_install": False,
-    "installable": False,
+    "installable": True,
+
+    "demo_title": "Sync POS orders across multiple sessions",
+    "demo_addons": [
+        "pos_disable_payment",
+        "pos_multi_session_sync",
+        "pos_multi_session_restaurant",
+    ],
+    "demo_addons_hidden": [
+    ],
+    "demo_url": "pos-multi-session",
+    "demo_summary": "Use multiple POSes for handling orders",
+    "demo_images": [
+        "images/pos-multi-session.png",
+    ]
 }