+++ pos_multi_session/external_tests/tests/__init__.py
@@ -1,2 +1,4 @@
-#from . import test_framework
+# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl.html).
+
+from . import test_framework
 from . import test_sync