+++ pos_multi_session/external_tests/tests/test_framework.py
@@ -1,4 +1,7 @@
 # -*- coding: utf-8 -*-
+# Copyright 2016 Ivan Yelizariev <https://it-projects.info/team/yelizariev>
+# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl.html).
+
 from .common import TestCommon
 
 
@@ -6,21 +9,26 @@ class TestSync(TestCommon):
     def test_10_sessions(self):
         """Check framework"""
         self.phantom_js_multi({
-            "admin": {
-                "url_path": "/pos/web?debug",
-                "login": "admin",
-            },
-            "demo": {
-                "url_path": "/pos/web",
-                "login": "demo",
-            }
+            "admin": {},
+            "demo": {},
         }, [
+            # check admin authentication
             {"session": "admin",
              "code": "$('.username:contains(Administrator)').length || console.log('error', 'Administrator label is not found')",
              },
+            # extra time for demo
             {"session": "demo",
+             "code": """
+                 mstest.wait(function(){
+                 }, 2000)
+            """,
+             },
+            # check demo authentication
+            {"session": "demo",
+             "screenshot": "test-framework-user",
              "code": "$('.username:contains(Demo)').length || console.log('error', 'Demo label is not found');"
              },
+            # check admin authentication
             {"session": "admin",
              "code": "$('.username:contains(Administrator)').length || console.log('error', 'Administrator label is not found')",
              },
@@ -31,10 +39,7 @@ class TestSync(TestCommon):
 
     def test_20_inject(self):
         self.phantom_js_multi({
-            "admin": {
-                "url_path": "/pos/web",
-                "login": "admin",
-            },
+            "admin": {},
         }, [
             {"session": "admin",
              "code": "testInject()?console.log('ok'):console.log('error', 'js in not injected')",