+++ pos_multi_session/static/src/js/pos_multi_session.js
@@ -1,3 +1,11 @@
+/* Copyright 2015-2016,2018 Ivan Yelizariev <https://it-projects.info/team/yelizariev>
+ * Copyright 2015-2016 Ilyas Rakhimkulov
+ * Copyright 2016 Gael Torrecillas
+ * Copyright 2016-2018 Dinar Gabbasov <https://it-projects.info/team/GabbasovDinar>
+ * Copyright 2017 Kolushov Alexandr <https://it-projects.info/team/KolushovAlexandr>
+ * Copyright 2017 David Arnold
+ * License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl.html). */
+
 odoo.define('pos_multi_session', function(require){
     var exports = {};
 
@@ -10,11 +18,46 @@ odoo.define('pos_multi_session', function(require){
     var chrome = require('point_of_sale.chrome');
     var longpolling = require('pos_longpolling');
     var Model = require('web.Model');
+    var PosBaseWidget = require('point_of_sale.BaseWidget');
+    var gui = require('point_of_sale.gui');
+    var framework = require('web.framework');
+
 
     var _t = core._t;
 
 
     screens.OrderWidget.include({
+        renderElement: function(scrollbottom){
+            var order = this.pos.get_order();
+            if (!order) {
+                return;
+            }
+            if (this.pos.debug) {
+                console.log("renderElement of OrderWidget");
+            }
+            this._super(scrollbottom);
+        },
+        bind_order_events: function() {
+            this._super();
+            var order = this.pos.get_order();
+            order.unbind('change:newLines', this.change_current_order, this);
+            order.bind('change:newLines', this.change_current_order, this);
+        },
+        change_current_order: function() {
+            if (this.pos.get_order()) {
+                this.renderElement('and_scroll_to_bottom');
+            }
+        },
+        orderline_add: function(){
+            var options = arguments && arguments[2]
+            ? arguments[2]
+            : {};
+            if (options.not_render) {
+                this.numpad_state.reset();
+            } else {
+                this._super();
+            }
+        },
         rerender_orderline: function(order_line){
             if (order_line.node && order_line.node.parentNode) {
                 return this._super(order_line);
@@ -28,6 +71,15 @@ odoo.define('pos_multi_session', function(require){
                 return this._super(order_line);
             }
         }
+    });
+    screens.set_fiscal_position_button.include({
+        button_click: function() {
+            var self = this;
+            this._super(event);
+            this.gui.current_popup.$(".selection-item").click(function(){
+                self.pos.get_order().trigger('change:sync');
+            });
+        },
     });
     screens.ReceiptScreenWidget.extend({
         finish_order: function() {
@@ -46,14 +98,33 @@ odoo.define('pos_multi_session', function(require){
     models.PosModel = models.PosModel.extend({
         initialize: function(){
             var self = this;
+            var ms_model = {
+                model: 'pos.multi_session',
+                fields: ['run_ID', 'multi_session_active'],
+                domain: function(){
+                    return [['id', '=', self.config.multi_session_id[0]]];
+                },
+                loaded: function(me, current_session) {
+                    self.multi_session_active = current_session[0].multi_session_active;
+                    if (self.multi_session_active) {
+                        self.multi_session_run_ID = current_session[0].run_ID;
+                    }
+            }};
+            this.models.splice(
+                this.models.indexOf(_.find(this.models, function(model){
+                    return model.model === 'pos.config';
+                })) + 1, 0, ms_model);
             PosModelSuper.prototype.initialize.apply(this, arguments);
             if (!this.message_ID) {
-                this.message_ID = 1;
+                this.message_ID = 0;
             }
             this.multi_session = false;
             this.ms_syncing_in_progress = false;
+            if (this.getUrlParameter('stringify-logs') === "1") {
+                this.stringify_logs = true;
+            }
             this.ready.then(function () {
-                if (!self.config.multi_session_id){
+                if (!self.multi_session_active){
                     return;
                 }
                 self.get('orders').bind('remove', function(order, collection, options){
@@ -68,11 +139,58 @@ odoo.define('pos_multi_session', function(require){
                     }
                     order.ms_remove_order();
                 });
-                self.multi_session = new exports.MultiSession(self);
                 var channel_name = "pos.multi_session";
                 var callback = self.ms_on_update;
-                self.add_channel(channel_name, callback, self);
+                self.bus.add_channel_callback(channel_name, callback, self);
+                if (self.config.sync_server){
+                    callback = self.ms_on_update;
+                    self.add_bus('sync_server', self.config.sync_server);
+                    self.get_bus('sync_server').add_channel_callback(channel_name, callback, self);
+                    self.sync_bus = self.get_bus('sync_server');
+                    self.get_bus('sync_server').start();
+                } else {
+                    self.sync_bus = self.get_bus();
+                    if (!self.config.autostart_longpolling) {
+                        self.sync_bus.start();
+                    }
+                }
+                self.multi_session.bind_ms_connection_events();
             });
+        },
+        after_load_server_data: function() {
+            var self = this;
+            var res = PosModelSuper.prototype.after_load_server_data.apply(this, arguments);
+            if (!this.multi_session_active){
+                return res;
+            }
+            this.multi_session = new exports.MultiSession(self);
+            var done = new $.Deferred();
+
+            $.when(res).then(function() {
+                var progress = (self.models.length - 0.5) / self.models.length;
+                self.chrome.loading_message(_t('Sync Orders'), progress);
+
+                return self.multi_session.request_sync_all({'immediate_rerendering': true}).then(function() {
+                    done.resolve();
+                });
+            });
+            return done;
+        },
+        getUrlParameter: function(sParam) {
+            var sPageURL = decodeURIComponent(window.location.search.substring(1)),
+                sURLVariables = sPageURL.split('&'),
+                sParameterName = '',
+                i = 0;
+
+            for (i = 0; i < sURLVariables.length; i++) {
+                sParameterName = sURLVariables[i].split('=');
+
+                if (sParameterName[0] === sParam) {
+                    return typeof sParameterName[1] === 'undefined'
+                           ? true
+                           : sParameterName[1];
+                }
+            }
         },
         ms_my_info: function(){
             var user = this.cashier || this.user;
@@ -109,36 +227,42 @@ odoo.define('pos_multi_session', function(require){
             return PosModelSuper.prototype.on_removed_order.apply(this, arguments);
         },
         ms_on_update: function(message, sync_all){
-            this.ms_syncing_in_progress = true; // don't broadcast updates made from this message
+            // don't broadcast updates made from this message
+            this.ms_syncing_in_progress = true;
             var error = false;
             var self = this;
             var data = '';
             var action = '';
             try{
                 if (this.debug){
-                    console.log('MS', this.config.name, 'on_update:', JSON.stringify(message));
+                    var logs = message;
+                    if (this.stringify_logs) {
+                        logs = JSON.stringify(logs);
+                    }
+                    console.log('MS', this.config.name, 'on_update:', logs);
                 }
                 action = message.action;
                 data = message.data || {};
                 var order = false;
                 if (data.uid){
-                    order = this.get('orders').find(function(order){
-                        return order.uid == data.uid;
+                    order = this.get('orders').find(function(item){
+                        return item.uid === data.uid;
                     });
                 }
                 if (sync_all) {
                     this.message_ID = data.message_ID;
                     this.ms_do_update(order, data);
                 } else {
-                    if (self.message_ID + 1 != data.message_ID)
-                        self.multi_session.request_sync_all();
-                    else
+                    if (self.message_ID + 1 === data.message_ID) {
                         self.message_ID = data.message_ID;
-                    if (order && action == 'remove_order') {
-                        order.destroy({'reason': 'abandon'});
+                    } else {
+                        self.multi_session.request_sync_all();
                     }
-                    else if (action == 'update_order')
+                    if (order && action === 'remove_order') {
+                        order.destroy({'reason': 'abandon'});
+                    } else if (action === 'update_order'){
                         this.ms_do_update(order, data);
+                    }
                 }
             }catch(err){
                 error = err;
@@ -153,7 +277,7 @@ odoo.define('pos_multi_session', function(require){
             if (!current_order) {
                 return;
             }
-            is_frozen = !current_order.ms_replace_empty_order;
+            var is_frozen = !current_order.ms_replace_empty_order;
             if (this.config.multi_session_replace_empty_order && current_order.new_order && !is_frozen) {
                 current_order.destroy({'reason': 'abandon'});
             } else if (is_frozen || !current_order.new_order || !this.config.multi_session_deactivate_empty_order) {
@@ -163,106 +287,186 @@ odoo.define('pos_multi_session', function(require){
         },
         ms_create_order: function(options){
             options = _.extend({pos: this}, options || {});
-            return new models.Order({}, options);
+            var order = new models.Order({}, options);
+            // init_locked blocks execution of save_to_db
+            // the order is unlocked at the end of ms_do_update
+            order.init_locked = true;
+            return order;
         },
         ms_do_update: function(order, data){
             var pos = this;
             this.pos_session.order_ID = data.sequence_number;
-            if (!order){
-                var create_new_order = pos.config.multi_session_accept_incoming_orders || !(data.ms_info && data.ms_info.created.user.id != pos.ms_my_info().user.id);
+            if (order){
+                // init_locked blocks execution of save_to_db
+                order.init_locked = true;
+                order.apply_ms_data(data);
+            } else {
+                var create_new_order = pos.config.multi_session_accept_incoming_orders || !(data.ms_info && data.ms_info.created.user.id !== pos.ms_my_info().user.id);
                 if (!create_new_order){
                     return;
                 }
-                json = {
+                var json = {
                     sequence_number: data.sequence_number,
                     uid: data.uid,
-                    pos_session_id: this.pos_session.id,
+                    run_ID: data.run_ID,
                     statement_ids: false,
                     lines: false,
                     multiprint_resume: data.multiprint_resume,
                     new_order: false,
                     order_on_server: true,
+                    table_id: data.table_id,
+                    floors_id: data.floor_id
                 };
                 order = this.ms_create_order({ms_info:data.ms_info, revision_ID:data.revision_ID, data:data, json:json});
                 var current_order = this.get_order();
                 this.get('orders').add(order);
                 this.ms_on_add_order(current_order);
-            } else {
-                order.apply_ms_data(data);
             }
             var not_found = order.orderlines.map(function(r){
                 return r.uid;
             });
-            if(data.partner_id !== false)
-            {
+            if(data.partner_id === false) {
+                order.set_client(null);
+            } else {
                 var client = order.pos.db.get_partner_by_id(data.partner_id);
-                if(!client)
-                {
+                if(!client) {
 
-                    $.when(this.load_new_partners_by_id(data.partner_id))
-                                    .then(function(client){client = order.pos.db.get_partner_by_id(data.partner_id);
-                             order.set_client(client);},function(){});
+                    $.when(this.load_new_partners_by_id(data.partner_id)).
+                    then(function(new_client){
+                        new_client = order.pos.db.get_partner_by_id(data.partner_id);
+                        order.set_client(new_client);
+                    }, function(){
+                       // do nothing.
+                    });
                 }
                 order.set_client(client);
             }
-            else
-            {
-                order.set_client(null);
-            }
 
+            var added_new_lines = false;
             _.each(data.lines, function(dline){
                 dline = dline[2];
                 var line = order.orderlines.find(function(r){
-                    return dline.uid == r.uid;
+                    return dline.uid === r.uid;
                 });
                 not_found = _.without(not_found, dline.uid);
                 var product = pos.db.get_product_by_id(dline.product_id);
                 if (!line){
                     line = new models.Orderline({}, {pos: pos, order: order, product: product});
                     line.uid = dline.uid;
+                    line.apply_ms_data(dline);
+                    order.orderlines.add(line, {'not_render': true});
+                    added_new_lines = true;
+                } else if (dline.is_changed) {
+                    line.apply_ms_data(dline);
                 }
-                line.apply_ms_data(dline);
-                order.orderlines.add(line);
             });
-
+            if (added_new_lines) {
+                order.trigger('change:newLines', order);
+            }
             _.each(not_found, function(uid){
                 var line = order.orderlines.find(function(r){
-                               return uid == r.uid;
+                               return uid === r.uid;
                            });
                 order.orderlines.remove(line);
             });
             order.order_on_server = true;
             order.new_order = false;
+            order.init_locked = false;
         },
         load_new_partners_by_id: function(partner_id){
-        var self = this;
-        var def  = new $.Deferred();
-        var client;
-        var fields = _.find(this.models,function(model){ return model.model === 'res.partner'; }).fields;
-        new Model('res.partner')
-            .query(fields)
-            .filter([['id','=',partner_id]])
-            .all({'timeout':3000, 'shadow': true})
-            .then(function(partners){
-                if (self.db.add_partners(partners)) {   // check if the partners we got were real updates
-                    def.resolve();
-                } else {
+            var self = this;
+            var def = new $.Deferred();
+            var fields = _.find(this.models,function(model){
+                return model.model === 'res.partner';
+            }).fields;
+            new Model('res.partner').
+                query(fields).
+                filter([['id','=',partner_id]]).
+                all({'timeout':3000, 'shadow': true}).
+                then(function(partners){
+                     // check if the partners we got were real updates
+                    if (self.db.add_partners(partners)) {
+                        def.resolve();
+                    } else {
+                        def.reject();
+                    }
+                }, function(err,event){
+                    if (err) {
+                        console.log(err);
+                    }
+                    event.preventDefault();
                     def.reject();
-                }
-            }, function(err,event){ event.preventDefault(); def.reject(); });
-        return def;
-    },
+                });
+            return def;
+        },
     });
 
     chrome.OrderSelectorWidget.include({
         init: function(parent,options) {
             this._super(parent,options);
+            this.saved_data = false;
             this.pos.get('orders').bind('change:update_new_order', this.renderElement,this );
         },
         destroy: function(){
             this.pos.get('orders').unbind('change:update_new_order', this.renderElement, this);
             this._super();
         },
+        renderElement: function(){
+            var self = this;
+            if (this.compare_data()) {
+                return false;
+            }
+            if (this.pos.debug) {
+                console.log("renderElement of OrderSelectorWidget");
+            }
+            this._super();
+            this.save_changes_data();
+        },
+        save_changes_data: function() {
+            var orders = this.pos.get_order_list();
+            var collection = [];
+            var current_order = this.pos.get_order();
+            var selected = false;
+            orders.forEach(function(order) {
+                if (current_order && current_order.uid === order.uid) {
+                    selected = true;
+                } else {
+                    selected = false;
+                }
+                collection.push({
+                    'sequence_number': order.sequence_number,
+                    'new_order': order.new_order,
+                    'selected': selected
+                });
+            });
+            this.saved_data = JSON.stringify(collection);
+        },
+        compare_data: function() {
+            var orders = this.pos.get_order_list();
+            var collection = [];
+            var current_order = this.pos.get_order();
+            var selected = false;
+            orders.forEach(function(order) {
+                if (current_order && current_order.uid === order.uid) {
+                    selected = true;
+                } else {
+                    selected = false;
+                }
+                collection.push({
+                    'sequence_number': order.sequence_number,
+                    'new_order': order.new_order,
+                    'selected': selected
+                });
+            });
+            return this.saved_data === JSON.stringify(collection);
+        }
+    });
+
+    gui.Gui.include({
+        _close: function() {
+            this.closing = true;
+            this._super();
+        }
     });
 
     var is_first_order = true;
@@ -277,17 +481,21 @@ odoo.define('pos_multi_session', function(require){
             }
 
             OrderSuper.prototype.initialize.apply(this, arguments);
-            if (!this.pos.config.multi_session_id){
+            if (!this.pos.multi_session_active){
                 this.new_order = false;
                 return;
             }
             this.ms_info = {};
-            this.revision_ID = options.revision_ID || 1;
-
+            if (!this.revision_ID){
+                this.revision_ID = options.revision_ID || 1;
+            }
             if (!_.isEmpty(options.ms_info)){
                 this.ms_info = options.ms_info;
-            } else if (this.pos.multi_session){
+            } else if (this.pos.multi_session_active){
                 this.ms_info.created = this.pos.ms_my_info();
+            }
+            if (!this.run_ID) {
+                this.run_ID = this.pos.multi_session_run_ID || 1;
             }
             this.ms_replace_empty_order = is_first_order;
             is_first_order = false;
@@ -314,12 +522,15 @@ odoo.define('pos_multi_session', function(require){
             OrderSuper.prototype.set_client.apply(this,arguments);
         },
         ms_check: function(){
-            if (! this.pos.multi_session )
+            if (! this.pos.multi_session ){
                 return;
-            if (this.pos.ms_syncing_in_progress)
+            }
+            if (this.pos.ms_syncing_in_progress){
                 return;
-            if (this.temporary)
+            }
+            if (this.temporary){
                 return;
+            }
             return true;
         },
         ms_update: function(){
@@ -332,11 +543,13 @@ odoo.define('pos_multi_session', function(require){
             } else {
                 this.trigger('change');
             }
-            if (!this.ms_check())
+            if (!this.ms_check()){
                 return;
-            if (this.ms_update_timeout)
+            }
+            if (this.ms_update_timeout){
                 // restart timeout
                 clearTimeout(this.ms_update_timeout);
+            }
             this.ms_update_timeout = setTimeout(
                 function(){
                     self.ms_update_timeout = false;
@@ -349,14 +562,43 @@ odoo.define('pos_multi_session', function(require){
             }
             this.ms_info = data.ms_info;
             this.revision_ID = data.revision_ID;
+            if (data.fiscal_position_id) {
+                this.set_fiscal_position(data.fiscal_position_id);
+            }
+        },
+        set_fiscal_position: function(id) {
+            this.fiscal_position = _.find(this.pos.fiscal_positions, function(fp) {
+                return fp.id === id;
+            });
         },
         ms_remove_order: function(){
-            if (!this.ms_check())
+            if (!this.ms_check()){
                 return;
-            this.do_ms_remove_order();
+            }
+            var self = this;
+            if (this.ms_update_timeout){
+                // restart timeout
+                clearTimeout(this.ms_update_timeout);
+            }
+            this.ms_update_timeout = setTimeout(function(){
+                self.ms_update_timeout = false;
+                self.do_ms_remove_order();
+            }, 0);
         },
         do_ms_remove_order: function(){
-            this.pos.multi_session.remove_order({'uid': this.uid, 'revision_ID': this.revision_ID});
+            var self = this;
+            if (this.enquied){
+                return;
+            }
+            var f = function(){
+                self.enquied=false;
+                return self.pos.multi_session.remove_order({'uid': self.uid, 'revision_ID': self.revision_ID}).done();
+            };
+            if (!this.pos.multi_session_active){
+                return;
+            }
+            this.enquied = true;
+            this.pos.multi_session.enque(f);
         },
         export_as_JSON: function(){
             var data = OrderSuper.prototype.export_as_JSON.apply(this, arguments);
@@ -364,26 +606,30 @@ odoo.define('pos_multi_session', function(require){
             data.revision_ID = this.revision_ID;
             data.new_order = this.new_order;
             data.order_on_server = this.order_on_server;
+            data.run_ID = this.run_ID;
             return data;
         },
         init_from_JSON: function(json) {
             this.new_order = json.new_order;
             this.order_on_server = json.order_on_server;
+            this.revision_ID = json.revision_ID;
+            this.run_ID = json.run_ID;
             OrderSuper.prototype.init_from_JSON.call(this, json);
         },
         do_ms_update: function(){
             var self = this;
-            if (this.enquied)
+            if (this.enquied){
                 return;
+            }
             var f = function(){
                 self.enquied=false;
                 var data = self.export_as_JSON();
                 return self.pos.multi_session.update(data).done(function(res){
                     self.order_on_server = true;
-                    if (res && res.action=="update_revision_ID") {
+                    if (res && res.action === "update_revision_ID") {
                         var server_revision_ID = res.revision_ID;
                         var order_ID = res.order_ID;
-                        if (order_ID && self.sequence_number != order_ID) {
+                        if (order_ID && self.sequence_number !== order_ID) {
                             self.sequence_number = order_ID;
                             // sequence number replace
                             self.pos.pos_session.order_ID = order_ID;
@@ -392,11 +638,13 @@ odoo.define('pos_multi_session', function(require){
                         }
                         if (server_revision_ID && server_revision_ID > self.revision_ID) {
                             self.revision_ID = server_revision_ID;
+                            self.save_to_db();
                         }
+
                     }
-                })
+                });
             };
-            if (!this.pos.config.multi_session_id){
+            if (!this.pos.multi_session_active){
                 return;
             }
             this.enquied = true;
@@ -414,9 +662,10 @@ odoo.define('pos_multi_session', function(require){
                 return;
             }
             this.uid = this.order.generate_unique_id() + '-' + this.id;
-            if (this.order.screen_data.screen === "splitbill")
+            if (this.order.screen_data.screen === "splitbill"){
                 // ignore new orderline from splitbill tool
                 return;
+            }
             if (this.order.ms_check()){
                 this.ms_info.created = this.order.pos.ms_my_info();
             }
@@ -425,7 +674,8 @@ odoo.define('pos_multi_session', function(require){
                     line.ms_info.changed = line.order.pos.ms_my_info();
                     line.order.ms_info.changed = line.order.pos.ms_my_info();
                     var order_lines = line.order.orderlines;
-                    order_lines.trigger('change', order_lines); // to rerender line
+                    // to rerender line
+                    order_lines.trigger('change', line);
                     line.order.trigger('change:sync');
                 }
             });
@@ -445,13 +695,13 @@ odoo.define('pos_multi_session', function(require){
                 OrderlineSuper.prototype.apply_ms_data.apply(this, arguments);
             }
             this.ms_info = data.ms_info || {};
-            if(data.qty !== undefined){
+            if(typeof data.qty !== "undefined" && data.qty !== this.quantity){
                 this.set_quantity(data.qty);
             }
-            if(data.price_unit !== undefined){
+            if(typeof data.price_unit !== "undefined" && data.price_unit !== this.price){
                 this.set_unit_price(data.price_unit);
             }
-            if(data.discount !== undefined){
+            if(typeof data.discount !== "undefined" && data.discount !== this.discount){
                 this.set_discount(data.discount);
             }
         },
@@ -476,30 +726,93 @@ odoo.define('pos_multi_session', function(require){
             this.order_ID = null;
             this.update_queue = $.when();
             this.func_queue = [];
-            this.pos.longpolling_connection.on("change:poll_connection", function(status){
+            this.on_syncing = false;
+        },
+        bind_ms_connection_events: function() {
+            var self = this;
+            this.pos.sync_bus.longpolling_connection.on("change:poll_connection", function(status){
+                if (self.pos.gui.closing) {
+                    return;
+                }
                 if (status) {
                     if (self.offline_sync_all_timer) {
                         clearInterval(self.offline_sync_all_timer);
                         self.offline_sync_all_timer = false;
                     }
                     self.request_sync_all();
-                } else {
-                    if (!self.offline_sync_all_timer) {
-                        self.no_connection_warning();
-                        self.start_offline_sync_timer();
-                        if (self.pos.debug){
-                            console.log('MultiSession initialize error');
-                        }
+                } else if (!self.offline_sync_all_timer) {
+                    self.no_connection_warning();
+                    self.start_offline_sync_timer();
+                    if (self.pos.debug){
+                        console.log('MultiSession initialize error');
                     }
                 }
             });
         },
-        request_sync_all: function(){
-            var data = {};
-            return this.send({'action': 'sync_all', data: data});
+        request_sync_all: function(options){
+            if (this.on_syncing) {
+                return;
+            }
+            options = options || {};
+            var self = this;
+            this.on_syncing = true;
+            var data = {run_ID: this.pos.multi_session_run_ID};
+            var message = {'action': 'sync_all', data: data};
+            if (options.uid) {
+                message.uid = options.uid;
+            }
+            return this.send(message, options).always(function(){
+                self.on_syncing = false;
+            });
+        },
+        sync_all: function(data, options) {
+            var server_orders_uid = [];
+            var self = this;
+            function delay(ms) {
+                var d = $.Deferred();
+                setTimeout(function(){
+                    d.resolve();
+                }, ms);
+                return d.promise();
+            }
+
+            this.q = $.when();
+            var done = $.Deferred();
+
+            options = options || {};
+
+            if (options.immediate_rerendering) {
+                data.orders.forEach(function (item) {
+                    self.pos.ms_on_update(item, true);
+                    server_orders_uid.push(item.data.uid);
+                });
+            } else {
+                data.orders.forEach(function (item) {
+                    self.q = self.q.then(function(){
+                        self.pos.ms_on_update(item, true);
+                        return delay(100);
+                    });
+                    server_orders_uid.push(item.data.uid);
+                });
+            }
+
+            this.pos.pos_session.order_ID = data.order_ID;
+
+            if (data.order_ID !== 0) {
+                this.pos.pos_session.sequence_number = data.order_ID;
+            }
+
+            if (!options.sync_current_order) {
+                this.destroy_removed_orders(server_orders_uid);
+            }
+            self.q.then(function() {
+                done.resolve();
+            });
+
+            return done;
         },
         remove_order: function(data){
-            this.send({action: 'remove_order', data: data});
+            return this.send({action: 'remove_order', data: data});
         },
         update: function(data){
             return this.send({action: 'update_order', data: data});
@@ -512,35 +825,48 @@ odoo.define('pos_multi_session', function(require){
                     var next = $.Deferred();
                     var func1 = self.func_queue.shift();
                     func1().always(function () {
-                        next.resolve()
+                        next.resolve();
                     });
                     return next;
                 }
-            })
+            });
         },
         _debug_send_number: 0,
-        send: function(message){
+        send: function(message, options){
+            options = options || {};
             var current_send_number = 0;
             if (this.pos.debug){
+                var logs = message;
+                if (this.pos.stringify_logs) {
+                    logs = JSON.stringify(logs);
+                }
                 current_send_number = this._debug_send_number++;
-                console.log('MS', this.pos.config.name, 'send #' + current_send_number +' :', JSON.stringify(message));
+                console.log('MS', this.pos.config.name, 'send #' + current_send_number +' :', logs);
             }
             var self = this;
             message.data.pos_id = this.pos.config.id;
             var send_it = function () {
-                return openerp.session.rpc("/pos_multi_session/update", {
+                var temp = self.pos.config.sync_server || '';
+                if (options.address) {
+                    temp = options.address.serv;
+                }
+                return openerp.session.rpc(temp + "/pos_multi_session_sync/update", {
                     multi_session_id: self.pos.config.multi_session_id[0],
-                    message: message
-                });
+                    message: message,
+                    dbname: session.db,
+                    user_ID: self.pos.user.id
+                },{timeout:30000});
             };
             return send_it().fail(function (error, e) {
                 if (self.pos.debug){
                     console.log('MS', self.pos.config.name, 'failed request #'+current_send_number+':', error.message);
                 }
-                if(error.message === 'XmlHttpRequestError ') {
+                if(error.data.type === "xhrerror" || (error.data.name && error.data.name.search('NotFound'))) {
                     self.client_online = false;
                     e.preventDefault();
-                    self.pos.longpolling_connection.network_is_off();
+                    if (self.pos.sync_bus) {
+                        self.pos.sync_bus.longpolling_connection.network_is_off();
+                    }
                     if (!self.offline_sync_all_timer) {
                         if (self.pos.debug){
                             console.log('send, return send_it error');
@@ -553,31 +879,35 @@ odoo.define('pos_multi_session', function(require){
                 }
             }).done(function(res){
                 if (self.pos.debug){
-                    console.log('MS', self.pos.config.name, 'response #'+current_send_number+':', JSON.stringify(res));
+                    var logs_res = res;
+                    if (self.pos.stringify_logs) {
+                        logs_res = JSON.stringify(logs_res);
+                    }
+                    console.log('MS', self.pos.config.name, 'response #'+current_send_number+':', logs_res);
                 }
-                var server_orders_uid = [];
                 self.client_online = true;
 
-                if (res.action == "revision_error") {
+                if (res.action === "revision_error") {
                     var warning_message = _t('There is a conflict during synchronization, try your action again');
                     self.warning(warning_message);
-                    self.request_sync_all();
+                    self.request_sync_all({'uid': res.order_uid});
                 }
-                if (res.action == 'sync_all') {
-                    res.orders.forEach(function (item) {
-                        self.pos.ms_on_update(item, true);
-                        server_orders_uid.push(item.data.uid);
-                    });
-                    self.pos.pos_session.order_ID = res.order_ID;
-
-                    if (res.order_ID != 0) {
-                        self.pos.pos_session.sequence_number = res.order_ID;
+                if (res.action === 'sync_all') {
+                    var sync_options = {};
+                    if (options.immediate_rerendering) {
+                        sync_options.immediate_rerendering = options.immediate_rerendering;
                     }
-                    self.destroy_removed_orders(server_orders_uid);
+                    if (message.uid) {
+                        sync_options.sync_current_order = true;
+                    }
+                    self.sync_all(res, sync_options);
                 }
                 if (self.offline_sync_all_timer) {
                     clearInterval(self.offline_sync_all_timer);
                     self.offline_sync_all_timer = false;
+                }
+                if (self.pos.sync_bus) {
+                    self.pos.sync_bus.longpolling_connection.network_is_on();
                 }
             });
         },
@@ -594,8 +924,8 @@ odoo.define('pos_multi_session', function(require){
             orders.forEach(function(item) {
                 var remove_order = server_orders_uid.indexOf(item.uid);
                 if (remove_order === -1) {
-                    var order = self.pos.get('orders').find(function (order) {
-                        return order.uid == item.uid;
+                    var order = self.pos.get('orders').find(function (ord) {
+                        return ord.uid === item.uid;
                     });
                     order.destroy({'reason': 'abandon'});
                 }
@@ -603,10 +933,13 @@ odoo.define('pos_multi_session', function(require){
             self.send_offline_orders();
         },
         warning: function(warning_message){
-            this.pos.chrome.gui.show_popup('error',{
-                'title': _t('Warning'),
-                'body': warning_message,
-            });
+            console.info('warning', warning_message);
+            if (_.keys(this.pos.gui.popup_instances).length) {
+                this.pos.chrome.gui.show_popup('error',{
+                    'title': _t('Warning'),
+                    'body': warning_message,
+                });
+            }
         },
         send_offline_orders: function() {
             var self = this;
@@ -621,7 +954,7 @@ odoo.define('pos_multi_session', function(require){
             var self = this;
             self.offline_sync_all_timer = setInterval(function(){
                 self.request_sync_all();
-            }, 5000);
+            }, 5000 + (Math.floor((Math.random()*10)+1)*1000));
         },
         no_connection_warning: function(){
             var warning_message = _t("No connection to the server. You can create new orders only. It is forbidden to modify existing orders.");