+++ pos_multi_session_restaurant/__manifest__.py
@@ -1,11 +1,16 @@
 # -*- coding: utf-8 -*-
+# Copyright 2017 Ilmir Karamov <https://it-projects.info/team/ilmir-k>
+# Copyright 2017-2018 Dinar Gabbasov <https://it-projects.info/team/GabbasovDinar>
+# Copyright 2018 Ivan Yelizariev <https://it-projects.info/team/yelizariev>
+# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl.html).
+
 {
     "name": """Sync restaurant orders""",
     "summary": """Staff get order details immediately after waiter taps on tablet""",
     "category": "Point of Sale",
-    "live_test_url": 'http://apps.it-projects.info/shop/product/pos-multi-session-restaurant?version=10.0',
+    "live_test_url": 'http://apps.it-projects.info/shop/product/pos-multi-session?version=10.0',
     "images": ['images/s2.png'],
-    "version": "2.1.0",
+    "version": "10.0.3.2.2",
     "application": False,
 
     "author": "IT-Projects LLC, Ivan Yelizariev",
@@ -16,7 +21,7 @@
     "currency": "EUR",
 
     "depends": [
-        "pos_restaurant",
+        "pos_restaurant_base",
         "pos_multi_session",
     ],
     "external_dependencies": {"python": [], "bin": []},
@@ -34,16 +39,5 @@
     "post_init_hook": None,
 
     "auto_install": True,
-    "installable": False,
-
-    "demo_title": "Sync restaurant orders",
-    "demo_addons": [
-    ],
-    "demo_addons_hidden": [
-    ],
-    "demo_url": "pos-multi-session-restaurant",
-    "demo_summary": "Staff get order details immediately after waiter taps on tablet",
-    "demo_images": [
-        "images/s2.png",
-    ]
+    "installable": True,
 }